requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==4.0.1
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.20.46
ecdsa==0.18.0
fastapi==0.89.1
filelock==3.8.0
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==22.0
passlib==1.7.4
pexpect==4.8.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pyrsistent==0.18.1
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
